+++ boot2docker/param.sh
@@ -13,7 +13,7 @@ LOGFILE="/var/lib/boot2docker/log/vagrant-param.log"
 # -------------------------------------------------
 if [ -f $BOOT2DOCKER_EXTENSION_PARAM ]; then
    echo "== PARAM execution... [$(date)] "
-   $BOOT2DOCKER_EXTENSION_PARAM "$1" "$2" >> $LOGFILE 2>&1
+   $BOOT2DOCKER_EXTENSION_PARAM "$1" "$2" 2>&1 | tee $LOGFILE
 else
    echo "== PARAM not found, try PROVISION before [$(date)] "
 fi